+++ SharedAssemblyInfo.cs
@@ -30,9 +30,9 @@ using System.Runtime.InteropServices;
 //      Build Number
 //      Revision (0 for stable, 1+ for beta)
 //
-[assembly: AssemblyVersion("1.1.0.4")]
-[assembly: AssemblyFileVersion("1.1.0.4")]
-[assembly: AssemblyInformationalVersion("1.1.0-beta.4")]
+[assembly: AssemblyVersion("1.1.0.9")]
+[assembly: AssemblyFileVersion("1.1.0.9")]
+[assembly: AssemblyInformationalVersion("1.1.0-alpha.9")]
 
 // Neutral Language
 [assembly: NeutralResourcesLanguage("en-US")]